+++ src/main/java/umc/product/domain/notice/adviser/NoticeAdviser.java
@@ -4,10 +4,10 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import umc.product.domain.member.entity.Member;
 import umc.product.domain.notice.converter.response.NoticeConverter;
-import umc.product.domain.notice.dto.response.NoticeCheckResponse;
-import umc.product.domain.notice.dto.response.NoticeDetailResponse;
-import umc.product.domain.notice.dto.response.NoticeResponse;
-import umc.product.domain.notice.dto.response.list.NoticeListResponse;
+import umc.product.domain.notice.dto.response.member.list.NoticeCheckResponse;
+import umc.product.domain.notice.dto.response.member.list.NoticeDetailResponse;
+import umc.product.domain.notice.dto.response.member.list.NoticeResponse;
+import umc.product.domain.notice.dto.response.member.list.NoticeListResponse;
 import umc.product.domain.notice.entity.Notice;
 import umc.product.domain.notice.entity.enums.NoticeTarget;
 import umc.product.domain.notice.service.NoticeService;
